compat: media: introduce API version number
Clients can now track changes to the API and continue to work on
platforms where not a certain API is not available rather than
simply crashing when a symbol isn't found.

Generally all targets should try to stay in sync with the upstream
releases we do for libhybris but doing proper landing of the same
changes for all platforms at the same gets really challanging so
we take this simple approach to let clients fallback or refuse
working when a certain API version is not available.